Make game borders properly styled
I have decided to continue with having the board be structured
with div elements that are styled through flexbox rather that structured as a
table. The borders of the game are now styled properly through the
addition of more classes (.vertical and .horizontal) in the appropriate
block div elements. Finally, the colour of the board was changed.