Create factory function logic for players
I also started expanding the JavaScript logic by working on the
gameBoard module, but this needs some rethinking. Finally, png images of
the x and o icons have been added and can be found in the player factory
function logic. The idea is that when a block on the board is clicked,
an onclick attribute will trigger a function that will retrieve the
player's icon and insert it into the block.