Work out logic for game board
I created a factory function named blackFactory that serves as the
template for the block objects. Each block object has an appropriate
block element, and a block status. I also created a gameboard module
that creates the block objects, stores them in an array, and returns
said array. I now want to create a function that can manipulate the
object status when the appropriate block is clicked, and see if my logic
holds weight.