Create logic to check and change block status
Now when a block is clicked, the status of a blocked is checked and can
be manipulated. This is done now through the changeBlockStatus function
that is triggered by an onclick attribute found in each block element.
I had to add additional classes for each block that represent their
positions in the gameBoardBlocks array, which helps access the block
object when clicked. Finally I simplified the block factory function.
It no longer takes in any arguments and only contains the status
attribute. There is no need for the block element attribute.