Create logic to add marker
Function created to add marker after status of block is changed.
Additionally, I changed the logic of the gameflow module to include just
an event listener, rather than a while loop with two event listeners.
The turn status has also been changed from being a declared variable, to
being stored in the player objects, and they get updated after the
status of the block is changed. There is an issue in the logic however
that only allows for one x and one o to be on the board at a time.