Fix issue of disappearing player markers
My previous logic for the addMarker function did not link the icons to
anything, so the x marker, for example, would be taken from the previous
block and applied to the newly selected block. To fix this, I added an
image element property to the block factory function. This way, newly
selected blocks would link the player marker to the appropriate block
object. Additionally, in the player factory function, I removed the
created image element, adn simply initialized a constant with the URL of
the player marker, which makes it easier to link it to the block
objects. I also removed the class names that represented the block
object numbers and edited the logic of the script to work with the id
name numbers instead. Finally, I updated the CSS styling and replaced
flex-grow for each block with flex-basis so that the space doesn't grow
when a player marker is inserted.