Turn all global functions into private functions
I originally designed this game not keeping in mind closure. Now, all
previously made functions have been placed in the player and block
factory functions. Now, all properties of the player and block objects
are private thanks to this and thanks to new getter functions. I
kept some comments in the player factory function that describe the
reasoning behind getters and setters because I want to be able to refer
to these explanations later.