Add dynamic styling to show who's turn it is
I added a separate module to style the tic tac toe page dynamically. For
this commit, I added one method to this module, which is a method that
adds a highlighted bar under the player who's turn it is. This method is
declared twice in the gameFlow module; once before the event listener,
and once inside the event listener after turns are switched.