Write logic for game over and reset button
I wrote multiple methods, that are stored in the proper modules and
objects, that describe the logic for a game over and the reset button.
The winning screens of each player, as well as the draw screen, are
written in the html and styled in the CSS, so that they can be turned on
and off easily through the various methods. Additionally, I fixed a bug
that allowed a player to click an occupied box, which changed turns
without a marker being placed. This was done by adding another condition
to the block event listener.